requirements.txt
openai
streamlit
gTTS
python-docx
easyocr
opencv-python-headless
Pillow
numpy
duckduckgo-search
pypdf
streamlit-mic-recorder